+++ wolfcrypt/src/wc_port.c
@@ -4265,9 +4265,9 @@ time_t stm32_hal_time(time_t *t1)
 
 #if (!defined(WOLFSSL_LEANPSK) && !defined(STRING_USER)) || \
     defined(USE_WOLF_STRNSTR)
-char* wolfSSL_strnstr(const char* s1, const char* s2, unsigned int n)
+char* wolfSSL_strnstr(const char* s1, const char* s2, size_t n)
 {
-    unsigned int s2_len = (unsigned int)XSTRLEN(s2);
+    size_t s2_len = XSTRLEN(s2);
 
     if (s2_len == 0)
         return (char *)(wc_ptr_t)s1;
